requirements.txt
pypdf>=4.0.0
Pillow>=10.0.0
youtube-transcript-api>=0.6.2
pytest>=7.0.0
